+++ CHANGELOG.md
@@ -9,15 +9,20 @@
 - form looks
 - time format
 - env ver を cljs から参照できるか？
-- CLJS repl
+- CLJS repl がつながらなくなった理由。
 - check your report URL を csrf と同様の手段で cljs に渡す。
-- (js/Date.) or (js/Date)
+
 
 ## 0.4.0-SNAPSHOT
 - :page-url "http://localhost:3001/" の導入。
   デベロップでは http-server を動かしとけ。
   プロダクションでは "https://hp.melt.kyutech.ac.jp/" になる。
 - upload 後の戻り先 -> /r/#/upload とした。
+- (js/Date.) or (js/Date)
+  (js/Date) == (str (js/Date.))
+- `check your report` in upload-page
+- `check your report` in home-page
+
 
 ## 0.3.3 - 2022-05-23
 - /api/upload production では PUBLIC_DIR 環境変数を定義すること。